Replacing old cli mechanism to be using picocli
