+++ rskj-core/src/main/java/co/rsk/cli/PicoCliToolRskContextAware.java
@@ -18,31 +18,19 @@
 package co.rsk.cli;
 
 import co.rsk.RskContext;
-import co.rsk.config.RskSystemProperties;
-import co.rsk.util.Factory;
-import co.rsk.util.NodeStopper;
-import org.ethereum.datasource.DbKind;
-import org.ethereum.datasource.KeyValueDataSource;
-import org.ethereum.util.FileUtil;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 import picocli.CommandLine;
 
 import javax.annotation.Nonnull;
-import java.io.File;
 import java.io.IOException;
-import java.lang.invoke.MethodHandles;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.Objects;
 import java.util.concurrent.Callable;
-import java.util.stream.Stream;
 
 public abstract class PicoCliToolRskContextAware extends CliToolRskContextAware implements Callable<Integer> {
+    protected RskContext ctx;
 
     @Override
     protected void onExecute(@Nonnull String[] args, @Nonnull RskContext ctx) throws IOException {
+        this.ctx = ctx;
+
         new CommandLine(this).execute(args);
     }
 }
