--- rskj-core/src/main/java/co/rsk/cli/tools/DbMigrateWithPicocli.java
@@ -1,52 +1,0 @@
-/*
- * This file is part of RskJ
- * Copyright (C) 2017 RSK Labs Ltd.
- *
- * This program is free software: you can redistribute it and/or modify
- * it under the terms of the GNU Lesser General Public License as published by
- * the Free Software Foundation, either version 3 of the License, or
- * (at your option) any later version.
- *
- * This program is distributed in the hope that it will be useful,
- * but WITHOUT ANY WARRANTY; without even the implied warranty of
- * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE. See the
- * GNU Lesser General Public License for more details.
- *
- * You should have received a copy of the GNU Lesser General Public License
- * along with this program. If not, see <http://www.gnu.org/licenses/>.
- */
-package co.rsk.cli.tools;
-
-import co.rsk.cli.PicoCliToolRskContextAware;
-import org.ethereum.datasource.DbKind;
-import org.ethereum.datasource.KeyValueDataSource;
-import org.ethereum.util.FileUtil;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
-import picocli.CommandLine;
-
-import java.io.File;
-import java.io.IOException;
-import java.lang.invoke.MethodHandles;
-import java.nio.file.Files;
-import java.nio.file.Path;
-import java.nio.file.Paths;
-import java.util.stream.Stream;
-
-@CommandLine.Command(name = "dbmigrate", mixinStandardHelpOptions = true, version = "dbmigrate 1.0",
-        description = "Migrates between different databases such as leveldb and rocksdb.")
-public class DbMigrateWithPicocli extends PicoCliToolRskContextAware {
-    @CommandLine.Parameters(index = "0", description = "The target db to migrate to. Example: leveldb, rocksdb ...")
-    private String targetdb;
-
-    public static void main(String[] args) {
-        create(MethodHandles.lookup().lookupClass()).execute(args);
-    }
-
-    @Override
-    public Integer call() throws IOException {
-        DbMigrate dbMigrate = new DbMigrate(this.ctx, this.stopper);
-        dbMigrate.execute(new String[]{targetdb});
-        return 0;
-    }
-}
